Hey Guys!
After two weeks of cooperative coding with Pilchenstein we came up with a toolbox called 

modules in a box (miab) which finaly allows this mod to.:

- LAND YOUR SPACESHIP ON THE PLANET:) Yay! It was the initial goal of the mod to realize 

that functionality. More about that on the main page. It is still far from perfect guys, 

but thats all the current lua allows us to do. Also 2 weeks of code are certainly buggy. 

Please report your finding per PM.
- Each tier features a landing pod "module" based on miab. A module is a whole ship or 

building compressed in a single item which can be deployed (see main page video)
- Starting at tier 5 you can build your own modules in game
- Objects have been tiered (small fin tier 1, medium fin tier 2 and so on)
- some custom crafting mats have been replaced by vanilla game version (i.e. PCB -> 

circuit board)
- "Shiptech" has been moved to the initial shiplocker loot to minimize things depending on 

player file
- rebalancing of object costs
- fixed a major bug in the life support which prevented it wrong working in bigger rooms
- added more race specific objects. I believe the first few floran parts are in
- ship station now has more filter options (icons at the bottom) for systems, hull parts, 

modules and so on.