Hello everyone :)!
- Furious Koala Support
- Starting Ships have been overhauled by HuggableCreep - they are a bit smaller so you have reason to improove them and look a lot nicer!
- Mining Pod. It is similar to the scoutship, very slow and has a large drill on the bottom side. Its tier5. We hope a lot, that this item is not imbalanced. Please comment on this in the discussion.
- Ressource System overhaul. We did mainly introduce a "metal press" which produces plates from ores. Those plate are then used as components for objects (i.e. fins) or to build blocks and tilesets. We hope we didnt overcomplicate it, but felt that a little bit more of production chain to the parts crafting could only improove it.
- numberous bugfixes as.:
-- Tilesetter will not not kill itself as it wont work when its anchored on tiles bachground
- more objects. Im not quite sure which ones- m0Ar :)

One more thing. As the starbound patch impact was big and i just had limited time to correct and test everything ... we exspect everything to work like a charm ... in fact there might be tons of new bugs in this build so please be understanding and post them to the discussions or pm me directly and we will fix the things asap.

Good luck and hf! :)