- added a "Ship Suit" Tech which can be equipped over the Suits page on the tech station.
It is basically just a host to allow me to run any kind of script on the player. I will also hook in all later suits there are in the game so that this allways works, no matter what suit you equip. Currently while equipping the Ship Suit Tech you can float in space (dry digging and jumping out of the ship). There will be other physics running based on that suit.

I didnt call it Spacesuit as i want that to be an Armor which you can put on. I believe theres one in Vanilla. Id like to use that in conjuctions with suffocation due to missing life support system and death in free space :).

- /techs folder renamed /tech for some compatibility with vanilla